[FIX] : 타입 및 프롭 정리
