+++ src/index.ts
@@ -1,4 +1,4 @@
 export { Table } from "./components/Table";
-export { SortableHeader } from "./components/Table";
-export type { SortableHeaderProps } from "./types";
+export { TableHeader } from "./components/TableHeader";
+export type { TableHeaderProps } from "./types";
 export type * from "./types";
