+++ src/main.tsx
@@ -1,7 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
+import { Table } from "./components/Table";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
-  <React.StrictMode></React.StrictMode>
+  <React.StrictMode>
+    <Table data={[]} columns={[]} storageKey="" />
+  </React.StrictMode>
 );
